description of the exercise , changes in the map generation and structures for the evaluation
